Move Json to another file
